+++ client/src/main/scala/drt/client/components/DateSelector.scala
@@ -4,10 +4,9 @@ import drt.client.logger.{Logger, LoggerFactory}
 import drt.client.services.JSDateConversions.SDate
 import drt.shared.SDateLike
 import japgolly.scalajs.react.component.Scala.Component
-import japgolly.scalajs.react.extra.Reusability
 import japgolly.scalajs.react.vdom.TagOf
 import japgolly.scalajs.react.vdom.html_<^._
-import japgolly.scalajs.react.{Callback, CallbackTo, CtorType, ReactEventFromInput, ScalaComponent}
+import japgolly.scalajs.react.{Callback, CallbackTo, CtorType, ReactEventFromInput, Reusability, ScalaComponent}
 import org.scalajs.dom.html.{Div, Select}
 
 import scala.scalajs.js.Date
@@ -23,7 +22,7 @@ object DateSelector {
 
   val today: SDateLike = SDate.now()
 
-  def daysInMonth(month: Int, year: Int): Int = new Date(year, month, 0).getDate()
+  def daysInMonth(month: Int, year: Int): Int = new Date(year, month, 0).getDate().toInt
 
   def formRow(label: String, xs: TagMod*): TagOf[Div] = {
     <.div(^.className := "form-group row",
